Remove docs for high level for now
