+++ docs/reference/bindings.md
@@ -1,0 +1,9 @@
+# `egg_smol.bindings`
+
+This modules contains bindings for the rust `egg-smol` library which are meant to be as close to the source as possible. This might result in not the most ergonomic API, if so, we can build higher level abstractions on top of it.
+
+```{eval-rst}
+.. automodule:: egg_smol.bindings
+   :members:
+   :undoc-members:
+```
