+++ docs/reference/high-level.md
@@ -1,0 +1,13 @@
+# `egg_smol`
+
+The top level module contains the high level API for using e-graphs in Python.
+
+The high level API is not documented yet, becuase adding supporting for our custom objects requires [a custom AutoDoc extension](https://www.sphinx-doc.org/en/master/development/tutorials/autodoc_ext.html#autodoc-ext-tutorial).
+
+<!--
+```{eval-rst}
+.. automodule:: egg_smol
+   :members:
+   :imported-members:
+   :special-members: __ne__
+``` -->
